Somewhat significant refactor
Refactor the DB schema. Rename workspace to space for brevity.
Everything is owned by a space except for identity. I'm still not a
super huge fan on the name "identity" but it represents a single email
within the system.

Created tables user, ticket, and assigneee to demonstrate the new schema
and the strengths of this approach. Since `space_id` is a part of every
primary key in every table owned by a space, we have really consistency
in ownership - it is impossible for a `user` in 1 space to be assigned
to a `ticket` in another space.

Added a few tests - more to come. Need a way to add users to spaces so
that we can excerise the special securiety properties of the schema
described above.